Sync

+++ libs/GameLogic/WorldObject/cave.h
@@ -11,7 +11,7 @@ class Cave : public Building
 protected:
     int             m_troopType;
     int             m_unitId;
-    float           m_spawnTimer;
+    double           m_spawnTimer;
     bool            m_dead;
 
     ShapeMarker     *m_spawnPoint;
@@ -20,7 +20,7 @@ public:
     Cave();
 
     bool Advance();
-    void Damage( float _damage );
+    void Damage( double _damage );
 };
