+++ libs/GameLogic/WorldObject/lasertrooper.h
@@ -11,14 +11,14 @@ public:
     Vector3 m_targetPos;
     Vector3 m_unitTargetPos;
 
-    float   m_victoryDance;
-
+    double   m_victoryDance;
+	
 public:
-
+    
     void Begin      ();
 
-    bool Advance    ( Unit *_unit );
-    void Render     ( float _predictionTime, int _teamId );
+    bool Advance    ( Unit *_unit );   
+    void Render     ( double _predictionTime, int _teamId );
 
     void AdvanceVictoryDance();
 };
