+++ libs/GameLogic/WorldObject/teleport.h
@@ -1,12 +1,12 @@
 #ifndef _included_teleport_h
 #define _included_teleport_h
 
-#include "vector3.h"
+#include "lib/vector3.h"
 
 #include "worldobject/building.h"
 
 class Entity;
-
+     
 struct TeleportMap
 {
     int m_teamId;
@@ -16,17 +16,17 @@ struct TeleportMap
 
 
 class Teleport : public Building
-{
+{   
 protected:
 
-    float       m_timeSync;
-    float       m_sendPeriod;                                   // Minimum time between sends
+    double       m_timeSync;
+    double       m_sendPeriod;                                   // Minimum time between sends
     static      LList<TeleportMap> m_teleportMap;               // Maps units going in onto units comingout
 
     ShapeMarker *m_entrance;
 
 protected:
-
+    
     void RenderSpirit   ( Vector3 const &_pos, int _teamId );
 
 public:
@@ -38,21 +38,21 @@ public:
     void SetShape       ( Shape *_shape );
 
     bool Advance        ();
-    void RenderAlphas   ( float predictionTime );
-
+    void RenderAlphas   ( double predictionTime );
+    
     void EnterTeleport  ( WorldObjectId _id, bool _relay=false );      // Relay=true means i've entered directly from another teleport
 
     bool IsInView       ();
 
     virtual bool        Connected();
     virtual bool        ReadyToSend ();
-
+    
     virtual bool        GetEntrance ( Vector3 &_pos, Vector3 &_front );
     virtual bool        GetExit     ( Vector3 &_pos, Vector3 &_front );
-
+    
     virtual Vector3     GetStartPoint();
     virtual Vector3     GetEndPoint();
-
+    
     virtual bool        UpdateEntityInTransit( Entity *_entity );      // Returns true (remove me) or false (still inside)
 
 };
